Allows for better control of column splitting
This now allows the user or the programmer of functions to specify if he
wants to split mixed type columns into separate numeric and categorical
columns.

Also implements a function that allows to coerce columns when imported
into R to numeric or categorical values based on information given in
the metadata.